@@ -3,51 +3,52 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build --prod",
+    "start": "ng serve --configuration development",
+    "build": "ng build --configuration production",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@sentry/angular": "^6.2.3",
+    "@angular/animations": "^20.2.1",
+    "@angular/common": "^20.2.1",
+    "@angular/compiler": "^20.2.1",
+    "@angular/core": "^20.2.1",
+    "@angular/forms": "^20.2.1",
+    "@angular/platform-browser": "^20.2.1",
+    "@angular/platform-browser-dynamic": "^20.2.1",
+    "@angular/router": "^20.2.1",
+    "@sentry/angular": "^10.7.0",
     "@sentry/tracing": "^6.2.3",
-    "bootstrap": "^4.4.1",
+    "bootstrap": "^5.3.8",
+    "compare-versions": "^3.6.0",
+    "emoji-regex": "^10.0.0",
     "jquery": "^3.5.0",
-    "popper.js": "^1.16.0",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
+    "rxjs": "~6.6.7",
+    "tslib": "^2.0.0",
     "websockhop": "^2.2.1",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular/cli": "^8.3.23",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "^20.2.1",
+    "@angular/cli": "^20.2.0",
+    "@angular/compiler-cli": "^20.2.1",
+    "@angular/language-service": "^20.2.1",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.2.1",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.1",
-    "protractor": "~5.4.0",
+    "@types/node": "^12.11.1",
+    "angular-eslint": "20.2.0",
+    "eslint": "^9.33.0",
+    "jasmine-core": "~5.6.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.4",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~5.8.3",
+    "typescript-eslint": "8.40.0"
   }
 }
