@@ -2,7 +2,8 @@
   "name": "keyman-status",
   "version": "1.0.0",
   "description": "Report on status of Keyman projects",
-  "main": "dist/code.js",
+  "main": "dist/server/code.js",
+  "type": "module",
   "scripts": {
     "build": "tsc",
     "start": "run-script-os",
@@ -12,23 +13,35 @@
   "author": "Marc Durdin",
   "license": "MIT",
   "dependencies": {
-    "@sentry/node": "^6.2.3",
-    "@sentry/tracing": "^6.2.3",
-    "@types/deep-equal": "^1.0.1",
-    "@types/parse-link-header": "^1.0.0",
+    "@octokit/core": "^5.2.0",
+    "@octokit/plugin-rest-endpoint-methods": "13.3.2-cjs.1",
+    "@octokit/rest": "^21.1.1",
+    "@sentry/node": "^9.8.0",
+    "@sentry/tracing": "^7.120.3",
+    "application-config-path": "^0.1.0",
     "deep-equal": "^2.0.5",
-    "express": "^4.17.1",
-    "parse-link-header": "^1.0.1",
+    "emoji-regex": "^10.2.1",
+    "express": "^5.1.0",
+    "parse-link-header": "^2.0.0",
+    "probot": "^12.0.0",
     "source-map-support": "^0.5.19",
-    "ws": "^7.4.4"
+    "ws": "^7.4.6"
   },
   "devDependencies": {
-    "@types/express": "^4.17.11",
+    "@octokit/types": "^12.3.0",
+    "@types/deep-equal": "^1.0.1",
+    "@types/express": "^5.0.1",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.1",
+    "@types/parse-link-header": "^1.0.0",
+    "@types/ws": "^8.18.1",
     "@typescript-eslint/eslint-plugin": "^4.18.0",
     "@typescript-eslint/parser": "^4.18.0",
     "eslint": "^7.22.0",
-    "run-script-os": "^1.1.1",
+    "mocha": "^10.2.0",
+    "run-script-os": "^1.1.6",
+    "ts-node": "^10.9.2",
     "tsc-watch": "^4.2.9",
-    "typescript": "^4.2.3"
+    "typescript": "^5.8.3"
   }
 }
