@@ -2,7 +2,8 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
     name: 'filterObjectByDate',
-    pure: false
+    pure: false,
+    standalone: false
 })
 export class FilterObjectByDatePipe implements PipeTransform {
     transform(items: any[], filter: Date): any {
@@ -14,6 +15,8 @@ export class FilterObjectByDatePipe implements PipeTransform {
         return items.filter(item => {
             let od = new Date(item.occurredAt);
             return od.getDate() == filter.getDate() && od.getMonth() == filter.getMonth() && od.getFullYear() == filter.getFullYear();
-        });
+        }).sort(
+            (a,b) => new Date(a.occurredAt).valueOf() - new Date(b.occurredAt).valueOf()
+        );
     }
 }
