@@ -1,48 +1,58 @@
-import { Component, OnInit, Input, ViewChild } from '@angular/core';
+import { Component, OnInit, OnChanges, Input } from '@angular/core';
+import { platformSentryIds } from '../../../../shared/platforms';
+import { PopupCoordinatorService } from '../popup-coordinator.service';
+import { PopupComponent } from '../popup/popup.component';
 import { siteSentryIds } from '../sites';
 import { escapeHtml } from '../utility/escapeHtml';
+import { VisibilityService } from '../visibility/visibility.service';
 
 @Component({
-  selector: 'app-sentry',
-  templateUrl: './sentry.component.html',
-  styleUrls: ['./sentry.component.css']
+    selector: 'app-sentry',
+    templateUrl: './sentry.component.html',
+    styleUrls: ['./sentry.component.css'],
+    standalone: false
 })
-export class SentryComponent implements OnInit {
+export class SentryComponent extends PopupComponent implements OnInit, OnChanges {
+  @Input() environment: string;
   @Input() platform?: string;
   @Input() site?: string;
   @Input() issues?: any;
+  @Input() mode?: string;
 
-  pinned: boolean = false;
+  env: any;
 
-  constructor() { }
+  constructor(popupCoordinator: PopupCoordinatorService, visibilityService: VisibilityService) {
+    super(popupCoordinator, visibilityService);
+  }
 
   ngOnInit() {
+    this.popupId = 'sentry-'+this.environment+'-'+(this.platform ? this.platform : this.site);
+    if(!this.gravityX) this.gravityX = 'right';
+    if(!this.gravityY) this.gravityY = 'bottom';
+    this.env = this.issues && this.issues[this.environment] ? this.issues[this.environment] : {
+      totalUsers: 0,
+      totalEvents: 0,
+      issues: []
+    };
+    super.ngOnInit();
+  }
+
+  issueTrackBy(index, item) {
+    return item.shortId;
+  }
+
+  ngOnChanges() {
+    this.env = this.issues && this.issues[this.environment] ? this.issues[this.environment] : {
+      totalUsers: 0,
+      totalEvents: 0,
+      issues: []
+    };
   }
 
   projectIndex(): number {
     if(!this.platform) return 0;
-    // TODO: consolidate with list in code.js
-    const map = { ...siteSentryIds, ...{
-      android:7,
-      developer:6,
-      ios:8,
-      linux:12,
-      mac:9,
-      web:11,
-      windows:5
-      }
-    };
+    const map = { ...siteSentryIds, ...platformSentryIds };
     return map[this.platform];
-  }
-
-  userCount() {
-    return this.issues ? this.issues.totalUsers : 0;
-  }
-  issueCount() {
-    return this.issues ? this.issues.issues.length : 0;
-  }
-  eventCount() {
-    return this.issues ? this.issues.totalEvents : 0;
   }
 
   // annotations are returned in the form "<a href=\"https://github.com/keymanapp/status.keyman.com/issues/88\">keymanapp/status.keyman.com#88</a>"
@@ -58,15 +68,11 @@ export class SentryComponent implements OnInit {
     return '';
   }
 
-  pin() {
-    this.pinned = !this.pinned;
-  }
-
   getSentryIssueText() {
-    if(!this.issues) return '';
+    if(!this.env) return '';
     const text =
       '<ul>' +
-      this.issues.issues.reduce(
+      this.env.issues.reduce(
         (text, node) => {
           return text + `<li>${escapeHtml(node.title)} (<a href='${node.permalink}'>${node.shortId}</a>)</li>\n`
         }, '') +
