@@ -1,36 +1,1 @@
-export const sites = [
-  'api.keyman.com',
-  'donate.keyman.com',
-  'downloads.keyman.com',
-  'help.keyman.com',
-  'keyman.com',
-  'keymanweb.com',
-  's.keyman.com',
-  'status.keyman.com', // ooh self referential!
-  'developer.keyman.com'
-];
-
-// TODO: this could be consolidated above
-export const siteSentryIds = {
-  'api.keyman.com': 4,
-  'developer.keyman.com': 14,
-  'donate.keyman.com': 15,
-  'downloads.keyman.com': 16,
-  'help.keyman.com': 2,
-  'keyman.com': 3,
-  'keymanweb.com': 10,
-  's.keyman.com': 17,
-  'status.keyman.com': 13
-};
-
-export const siteSentryNames = {
-  'api-keyman-com': 'api.keyman.com',
-  'developer-keyman-com': 'developer.keyman.com',
-  'donate-keyman-com': 'donate.keyman.com',
-  'downloads-keyman-com': 'downloads.keyman.com',
-  'help-keyman-com': 'help.keyman.com',
-  'keyman-com': 'keyman.com',
-  'keymanweb-com': 'keymanweb.com',
-  's-keyman-com': 's.keyman.com',
-  'status-keyman-com': 'status.keyman.com'
-};
+export * from '../../../shared/sites';
