@@ -1,38 +1,48 @@
-import * as https from "https";
+import * as https from "node:https";
+import * as http from "node:http";
 
-type resolver = (a: {data: string; res: any}) => void;
+type resolver = (a: {data: string; res: http.IncomingMessage}) => void;
 
-export default function httpget(hostname, path, headers?) {
-  return new Promise((resolve: resolver) => {
+export default function httpget(hostname, path, headers?, head?: boolean, httpOnly?: boolean) {
+  return new Promise((resolve: resolver, reject) => {
     const options: https.RequestOptions = {
       hostname: hostname,
-      port: 443,
+      port: httpOnly ? 80 : 443,
       path: path,
-      method: 'GET'
+      method: head ? 'HEAD' : 'GET'
     }
 
     if(headers) options.headers = headers;
 
     let chunk = '';
 
-    const req = https.request(options, res => {
-      if(res.statusCode != 200) {
-        console.error(`statusCode for ${hostname}${path}: ${res.statusCode}`);
-      }
+    try {
+      const req = (httpOnly ? http : https).request(options, res => {
+        if(res.statusCode != 200) {
+          const error = `statusCode for ${hostname}${path}: ${res.statusCode}`;
+          console.error(error);
+          //Sentry.captureMessage(`statusCode for ${hostname}${path}: ${res.statusCode}`);
+          reject(error);
+          return;
+        }
 
-      res.on('data', d => {
-        chunk += d;
+        res.on('data', d => {
+          chunk += d;
+          });
+
+        res.on('end', () => {
+          resolve({data: chunk, res: res});
         });
-
-      res.on('end', () => {
-        resolve({data: chunk, res: res});
       });
-    });
 
-    req.on('error', error => {
-      console.error(error);
-    });
+      req.on('error', error => {
+        reject(error);
+      });
 
-    req.end();
+      req.end();
+    } catch(e) {
+      console.log(e);
+      reject(e);
+    }
   });
 };
